requirements.txt
discord.py
python-dotenv
procyclingstats
unidecode
pycountry
matplotlib